дз семинар2 seminar105 long

+++ seminar105/Program.cs
@@ -6,12 +6,12 @@
 Console.Clear();
 
 Console.Write("Введите число: ");
-int num = int.Parse(Console.ReadLine());
+long num = long.Parse(Console.ReadLine());
 if (num < 100) 
 Console.WriteLine($"В введенном числе: {num} третья цифра отсутствует или введенное число отрицательное, введите новое число");
 else
     {
-    int element = num % 10; 
+    long element = num % 10; 
     while (num > 999)
         { 
         num = (num - element) / 10; element = num % 10;
